changelog: update versions and dependencies
This commit updates versions and dependencies for the project to include new packages and versions like floating-ui. It also adds the missing dependency '@vueuse/core' to package.json to resolve runtime issues for mods-module-docs. Additionally, it refactors code by extracting the Jagger component for reusability and readability. Lastly, it updates titles, content, and styles for better consistency and clarity in the project.