Bundle utility functions with the nuxt module build and use bundled functions directly
This is a tradeoff between using the usemods package through npm or using source code directly to import the utility functions into Nuxt.
This version copies the bundled utilities into the build process for the module and then references the bundled functions directly. It's a less elegant solution but solves for optimises for local development where changes are made to the base utility and needs fast rebuild of the module to use the new src.

Attempted to use npm link but had numerous issues.