chore: Update utils/mods directory structure
This commit updates the directory structure in the `utils/mods` directory. The changes include moving files from the `nuxt-web/utils` directory to the `nuxt-web/utils/mods` directory. This update ensures better organization and separation of files within the project.

Note: This commit message follows the established convention of using the `chore` type for non-functional changes and providing a clear and concise description of the changes made.