feat: Add content.config.ts for defining content collections
This commit adds a new file, content.config.ts, to the nuxt-web directory. The file defines the content collections for the project, specifically a collection of type 'page' with a source pattern of '**/*.md'. This will allow the project to handle markdown files as pages.

The commit also includes updates to the devDependencies in the package.json files of both nuxt-web and the root directory. The "@stylistic/eslint-plugin" dependency has been updated from version "^2.13.0" to "^3.0.1" in the root package.json, and the "@nuxt/eslint" dependency has been updated from version "^0.5.7" to "^1.0.0" in the nuxt-web package.json.

Lastly, the nuxt.config.ts file in the nuxt-web directory has been modified. The commented out sections related to "@nuxtjs/seo" and "@nuxtjs/sitemap" have been removed, and a new configuration option has been added under the "content" section. The "database" option has been set to true, enabling the use of a database for content storage.